Had to rename files as Kbuild was unhappy and not exporting symbols
